refactor: slipp 모듈에 있는 Jdbc Template 을 jdbc 모듈로 이동
